move infty.Type, infty.Instance to meta app; disable timezone support, cause we operate in utc
