+++ fixtures/currency_price_snapshots.json
