+++ fixtures/hour_price_snapshots.json
