+++ fixtures/instances.json
@@ -1,0 +1,1 @@
+[{"model": "meta.instance", "pk": 1, "fields": {"created_date": "2017-12-06T12:40:05.211Z", "updated_date": "2017-12-17T02:17:20.969Z", "languages": "[\"en\"]", "role": 1, "concept": 118, "identifiers": "Wefindx Foundation Company Limited By Guarantee", "description": ".:en\nFoundation registered in Ireland, 569345", "source": null, "data": {}}}, {"model": "meta.instance", "pk": 2, "fields": {"created_date": "2017-12-06T12:42:57.013Z", "updated_date": "2017-12-06T12:42:57.013Z", "languages": "[\"en\"]", "role": 0, "concept": 117, "identifiers": "WeFindX Model", "description": ".:en\nThe model of reality, specified at wefindx.net, more specifically, based on a hypothesis, that\r\n\r\nAgents set [Goals about the Needs] regarding (World: Agents, Assets, Events, Places, Topics) and pursue them by coming up with [Ideas, from which they derive Plans as networks of Steps] to realize them, and instantiate the Steps (prototypes for action) as actual Actions (world parametrization events).", "source": null, "data": {}}}]
