+++ infty/core/forms.py
@@ -1,28 +1,7 @@
 from django import forms
 
-from infty.core.models import Type, Instance, Topic, Comment
+from infty.core.models import Topic, Comment
 from infty.users.models import CryptoKeypair
-
-
-class TypeForm(forms.ModelForm):
-
-    class Meta:
-        model = Type
-        exclude = []
-
-
-class InstanceForm(forms.ModelForm):
-
-    role = forms.ChoiceField(
-        choices=Instance.ITEM_ROLES,
-        initial=Instance.THING
-    )
-
-    identifiers = forms.CharField()
-
-    class Meta:
-        model = Instance
-        exclude = []
 
 
 class TopicForm(forms.ModelForm):
